comments solved
